+++ drop-in/src/test/java/com/adyen/checkout/dropin/internal/ui/model/DropInParamsMapperTest.kt
@@ -15,10 +15,10 @@ import com.adyen.checkout.components.core.AnalyticsLevel
 import com.adyen.checkout.components.core.CheckoutConfiguration
 import com.adyen.checkout.components.core.internal.ui.model.AnalyticsParams
 import com.adyen.checkout.components.core.internal.ui.model.AnalyticsParamsLevel
+import com.adyen.checkout.components.core.internal.ui.model.SessionParams
 import com.adyen.checkout.core.Environment
 import com.adyen.checkout.dropin.DropInConfiguration
 import com.adyen.checkout.dropin.dropIn
-import com.adyen.checkout.sessions.core.internal.data.model.SessionDetails
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
@@ -115,22 +115,52 @@ internal class DropInParamsMapperTest {
     ) {
         val testConfiguration = createCheckoutConfiguration(configurationValue)
 
-        val sessionDetails = SessionDetails(
-            id = "id",
-            sessionData = "session_data",
+        val sessionParams = SessionParams(
+            enableStoreDetails = null,
+            installmentConfiguration = null,
             amount = sessionsValue,
-            expiresAt = "",
             returnUrl = null,
-            sessionSetupConfiguration = null,
+            shopperLocale = null,
         )
 
         val params = dropInParamsMapper.mapToParams(
             checkoutConfiguration = testConfiguration,
             deviceLocale = DEVICE_LOCALE,
-            sessionDetails = sessionDetails,
+            sessionParams = sessionParams,
         )
 
         val expected = getDropInParams(amount = expectedValue)
+
+        assertEquals(expected, params)
+    }
+
+    @ParameterizedTest
+    @MethodSource("shopperLocaleSource")
+    fun `shopper locale should match value set in configuration then sessions then device locale`(
+        configurationValue: Locale?,
+        sessionsValue: Locale?,
+        deviceLocaleValue: Locale,
+        expectedValue: Locale,
+    ) {
+        val configuration = createCheckoutConfiguration(shopperLocale = configurationValue)
+
+        val sessionParams = SessionParams(
+            enableStoreDetails = null,
+            installmentConfiguration = null,
+            amount = null,
+            returnUrl = "",
+            shopperLocale = sessionsValue,
+        )
+
+        val params = dropInParamsMapper.mapToParams(
+            checkoutConfiguration = configuration,
+            deviceLocale = deviceLocaleValue,
+            sessionParams = sessionParams,
+        )
+
+        val expected = getDropInParams(
+            shopperLocale = expectedValue,
+        )
 
         assertEquals(expected, params)
     }
@@ -187,6 +217,15 @@ internal class DropInParamsMapperTest {
             arguments(Amount("EUR", 100), null, Amount("EUR", 100)),
             arguments(null, Amount("CAD", 300), Amount("CAD", 300)),
             arguments(null, null, null),
+        )
+
+        @JvmStatic
+        fun shopperLocaleSource() = listOf(
+            // configurationValue, sessionsValue, deviceLocaleValue, expectedValue
+            arguments(null, null, Locale.US, Locale.US),
+            arguments(Locale.GERMAN, null, Locale.US, Locale.GERMAN),
+            arguments(null, Locale.CHINESE, Locale.US, Locale.CHINESE),
+            arguments(Locale.GERMAN, Locale.CHINESE, Locale.US, Locale.GERMAN),
         )
     }
 }